Remove duplicate versions of tests
