+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Bugzilla28570.cs
@@ -1,29 +1,31 @@
-﻿#if ANDROID
+﻿#if !ANDROID
 using NUnit.Framework;
 using UITest.Appium;
 using UITest.Core;
 
 namespace Microsoft.Maui.TestCases.Tests.Issues
 {
-	public class Bugzilla28570 : _IssuesUITest
+	[Category(UITestCategories.ScrollView)]
+	[Category(UITestCategories.Compatibility)]
+	public class Bugzilla28570UITests : _IssuesUITest
 	{
-		public Bugzilla28570(TestDevice testDevice) : base(testDevice)
+		public Bugzilla28570UITests(TestDevice device)
+			: base(device)
 		{
 		}
 
 		public override string Issue => "https://bugzilla.xamarin.com/show_bug.cgi?id=28570";
 
+		// Bugzilla28570 (src\Compatibility\ControlGallery\src\Issues.Shared\Bugzilla28570.cs)
 		[Test]
-		[Category(UITestCategories.Navigation)]
-		[Category(UITestCategories.Compatibility)]
-		[FailsOnAndroid("Scroll to end not working.")]
+		[FailsOnAndroid("This test is failing, likely due to product issue")]
 		public void Bugzilla28570Test()
 		{
-			App.WaitForElement("Tap");
-			App.Screenshot("At test page");
+			App.WaitForElement ("Tap");
+			App.Screenshot ("At test page");
 			App.Tap("Tap");
 
-			App.WaitForElement("28570Target");
+			App.WaitForElement ("28570Target");
 		}
 	}
 }
