+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/Bugzilla44461.cs
@@ -1,11 +1,13 @@
-﻿using NUnit.Framework;
+﻿#if ANDROID
+using NUnit.Framework;
 using NUnit.Framework.Legacy;
 using UITest.Appium;
 using UITest.Core;
 
-namespace Microsoft.Maui.TestCases.Tests
+namespace Microsoft.Maui.TestCases.Tests.Issues
 {
 	[Category(UITestCategories.ScrollView)]
+	[Category(UITestCategories.Compatibility)]
 	public class Bugzilla44461UITests : _IssuesUITest
 	{
 		public Bugzilla44461UITests(TestDevice device)
@@ -39,3 +41,4 @@ namespace Microsoft.Maui.TestCases.Tests
 		}
 	}
 }
+#endif
