+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/ScrollChangeOrientation.cs
@@ -3,7 +3,7 @@ using NUnit.Framework;
 using UITest.Appium;
 using UITest.Core;
 
-namespace Microsoft.Maui.TestCases.Tests
+namespace Microsoft.Maui.TestCases.Tests.Issues
 {
 	[Category(UITestCategories.ScrollView)]
 	public class ScrollChangeOrientationUITests : _IssuesUITest
