+++ src/Controls/tests/TestCases.Shared.Tests/Tests/Issues/ScrollViewNoContent.cs
@@ -2,7 +2,7 @@
 using UITest.Appium;
 using UITest.Core;
 
-namespace Microsoft.Maui.TestCases.Tests
+namespace Microsoft.Maui.TestCases.Tests.Issues
 {
 	[Category(UITestCategories.ScrollView)]
 	public class ScrollViewNoContentUITests : _IssuesUITest
